Add files via upload
subidos archivos actualizados

+++ listas.c
@@ -164,19 +164,19 @@ void ordenarpol (int * A, int * B , int N, int arreglo[][2], int * C){
 }
 
 /*****
-* float evaluar
+* double evaluar
 ******
 * evalua el polinomio a traves del algoritmo de horner, segun el valor pedido
 ******
 * Input:
 * tLista lista : lista enlazada que contiene a todos los polinomios
 * int posicion : posicion del polinomio que se quiere evaluar
-* float evaluado : el valor al cual se quiere evaluar el polinomio de la posicion determinada
+* double evaluado : el valor al cual se quiere evaluar el polinomio de la posicion determinada
 ******
 * Returns:
-* float, resultado numerico de la evaluacion del polinomio
+* double, resultado numerico de la evaluacion del polinomio
 *****/
-float evaluar(tLista lista, int posicion, float evaluado){
+double evaluar(tLista lista, int posicion, double evaluado){
     pol elemento = movetopost(lista, posicion);
     int exponen[elemento.tam];
     int exponen2[elemento.tam];
@@ -189,9 +189,9 @@ float evaluar(tLista lista, int posicion, float evaluado){
     }
     int elem[elemento.tam][2];
     ordenarpol(exponen2, coee, elemento.tam, elem, exponen);
-    float raiz = 0;
-    float coef = elem[0][1];
-    float poli = coef + raiz;
+    double raiz = 0;
+    double coef = elem[0][1];
+    double poli = coef + raiz;
     int i;
     for(i=1;i<elemento.tam;i++){
         int a = elem[i-1][0]-elem[i][0];
@@ -225,7 +225,7 @@ float evaluar(tLista lista, int posicion, float evaluado){
 * Returns:
 * int, retorna el valor del coeficiente pedido, si no existe el monomio del coeficiente, retorna 0
 *****/
-int coeficiente(tLista lista, int pos, float expo){
+int coeficiente(tLista lista, int pos, double expo){
     pol elemento = movetopost(lista, pos);
     int i;
     for(i=0;i<elemento.tam;i++){
@@ -261,20 +261,20 @@ int main(){
     char pp1[]="EVALUAR";
     char pp2[]="COEFICIENTE";
     int posleido;
-    float valorleido;
+    double valorleido;
     FILE *salida;
     salida=fopen("salidaPolinomio.txt","w");                                         // crea el archivo de salida
-    fscanf(archivo,"%s %d %f",funcion, &posleido, &valorleido);                      // lee archivo y almacena los valores en las valiares asignadas
+    fscanf(archivo,"%s %d %lf",funcion, &posleido, &valorleido);                      // lee archivo y almacena los valores en las valiares asignadas
     while(!feof(archivo)){                                                           // comprobar que no se ha llegado al EOF
         if (strncmp(funcion,pp1,7)==0){                                              // verifica si lo que se pide es EVALUAR
-            float resul = evaluar(lista_pol, posleido, valorleido);                  // evalua segun lo pedido
-            fprintf(salida,"%.6f\n",resul);                                          // escribe el resultado en el texto de salida
+            double resul = evaluar(lista_pol, posleido, valorleido);                  // evalua segun lo pedido
+            fprintf(salida,"%.6lf\n",resul);                                          // escribe el resultado en el texto de salida
         }
         else if(strncmp(funcion,pp2,11)==0){                                         // verifica si lo que se pide es COEFICIENTE
             int resul = coeficiente(lista_pol, posleido, valorleido);                // aplica la funcion segun lo pedido
             fprintf(salida,"%d\n",resul);                                            // escribe el resultado en el texto de salida
         }
-        fscanf(archivo,"%s %d %f",funcion, &posleido, &valorleido);                  // lee la siguiente linea del archivo
+        fscanf(archivo,"%s %d %lf",funcion, &posleido, &valorleido);                  // lee la siguiente linea del archivo
     }
     fclose(salida);
     free_linked(lista_pol.head);
